add method to join event with invitation code

+++ src/main/java/com/example/eowa/service/EventService.java
@@ -1,6 +1,7 @@
 package com.example.eowa.service;
 
 import com.example.eowa.exceptions.CalendarExceptions.CalendarException;
+import com.example.eowa.exceptions.authenticationExceptions.InvalidInvitationCodeException;
 import com.example.eowa.model.Calendar;
 import com.example.eowa.model.Event;
 import com.example.eowa.model.Opinion;
@@ -61,15 +62,15 @@ public class EventService {
         calendarService.setUnavailableHours(calendar, serialNumbers);
     }
 
-    public void setUnavailableHoursDaily(long id, Set<Integer> hourNumbers){
+    private void setUnavailableHoursDaily(long id, Set<Integer> hourNumbers){
         setUnavailableHoursPeriodically(id,hourNumbers, CalendarService.Period.DAILY);
     }
 
-    public void setUnavailableHoursWeekly(long id, Set<Integer> hourNumbers){
+    private void setUnavailableHoursWeekly(long id, Set<Integer> hourNumbers){
         setUnavailableHoursPeriodically(id,hourNumbers, CalendarService.Period.WEEKLY);
     }
 
-    private void setUnavailableHoursPeriodically(long id, Set<Integer> hourNumbers, int period){
+    public void setUnavailableHoursPeriodically(long id, Set<Integer> hourNumbers, int period){
         Calendar calendar = getEventById(id).getCalendar();
         if(calendar == null){
             return;
@@ -79,16 +80,33 @@ public class EventService {
 
     public void setUserOpinion(long id, Set<Integer> hourSerials, User user, Opinion.UserOpinion userOpinion){
         Calendar calendar = getEventById(id).getCalendar();
+        if(calendar == null){
+            return;
+        }
         calendarService.setUserOpinion(calendar,user,hourSerials,userOpinion);
     }
 
     public void removeUserOpinion(long id, Set<Integer> hourSerials, User user){
         Calendar calendar = getEventById(id).getCalendar();
+        if(calendar == null){
+            return;
+        }
         calendarService.removeUserOpinion(calendar,user,hourSerials);
     }
 
     public Event getEventById(Long id) {
         return eventRepository.findById(id).orElse(null);
+    }
+
+    public Event joinEventWithInvitationCode(User user, String invitationCode) throws InvalidInvitationCodeException {
+        Event event = eventRepository.findEventByInvitationCode(invitationCode);
+        if(event == null){
+            throw new InvalidInvitationCodeException();
+        }
+        if(!event.getParticipants().contains(user)){
+            event.addParticipant(user);
+        }
+        return event;
     }
 
     public Event updateEvent(Event event) {
